# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRI ANH EDU API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformStats {
  students: number;
  exams: number;
  questions: number;
  satisfactionRate: number;
}

export interface Course {
  id: number;
  title: string;
  teacherName: string;
  teacherAvatar: string;
  rating: number;
  studentsCount: number;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  thumbnail: string;
  level: string;
  /** @nullable */
  lessonsCount?: number | null;
  /** @nullable */
  durationHours?: number | null;
}

export interface CurriculumSection {
  title: string;
  lessons: string[];
}

export interface CourseDetail {
  id: number;
  title: string;
  teacherName: string;
  teacherAvatar: string;
  rating: number;
  studentsCount: number;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  thumbnail: string;
  level: string;
  description: string;
  whatYouLearn: string[];
  curriculum: CurriculumSection[];
  /** @nullable */
  lessonsCount?: number | null;
  /** @nullable */
  durationHours?: number | null;
}

export interface Exam {
  id: number;
  title: string;
  type: string;
  subject: string;
  difficulty: string;
  questionsCount: number;
  durationMinutes: number;
  attemptCount: number;
  isFree: boolean;
  /** @nullable */
  year?: number | null;
}

export interface Question {
  id: number;
  text: string;
  options: string[];
  correctOptionIndex: number;
  /** @nullable */
  explanation?: string | null;
}

export interface ExamDetail {
  id: number;
  title: string;
  type: string;
  subject: string;
  difficulty: string;
  questionsCount: number;
  durationMinutes: number;
  attemptCount: number;
  isFree: boolean;
  description: string;
  /** @nullable */
  year?: number | null;
  questions: Question[];
}

export interface Teacher {
  id: number;
  name: string;
  avatar: string;
  specialization: string;
  experienceYears: number;
  studentsCount: number;
  bio: string;
  /** @nullable */
  rating?: number | null;
}

export interface TeacherDetail {
  id: number;
  name: string;
  avatar: string;
  specialization: string;
  experienceYears: number;
  studentsCount: number;
  bio: string;
  /** @nullable */
  rating?: number | null;
  courses: Course[];
}

export interface BlogPost {
  id: number;
  title: string;
  category: string;
  readingTimeMinutes: number;
  thumbnail: string;
  publishedAt: string;
  excerpt: string;
}

export interface BlogPostDetail {
  id: number;
  title: string;
  category: string;
  readingTimeMinutes: number;
  thumbnail: string;
  publishedAt: string;
  excerpt: string;
  content: string;
  authorName: string;
  authorAvatar: string;
}

export interface Testimonial {
  id: number;
  studentName: string;
  studentAvatar: string;
  score: number;
  improvement: number;
  testimonialText: string;
  examType: string;
  /** @nullable */
  monthsStudied?: number | null;
}

export type ListCoursesParams = {
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListExamsParams = {
/**
 * @nullable
 */
type?: string | null;
/**
 * @nullable
 */
subject?: string | null;
/**
 * @nullable
 */
difficulty?: string | null;
};

export type ListBlogPostsParams = {
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
limit?: number | null;
};