# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRI ANH EDU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  category: string;
  readingTimeMinutes: number;
  thumbnail: string;
  publishedAt: string;
  excerpt: string;
}