# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRI ANH EDU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  title: string;
  teacherName: string;
  teacherAvatar: string;
  rating: number;
  studentsCount: number;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  thumbnail: string;
  level: string;
  /** @nullable */
  lessonsCount?: number | null;
  /** @nullable */
  durationHours?: number | null;
}