# lib/api-zod/src/generated/types/curriculumSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRI ANH EDU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CurriculumSection {
  title: string;
  lessons: string[];
}