# lib/api-zod/src/generated/types/exam.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRI ANH EDU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Exam {
  id: number;
  title: string;
  type: string;
  subject: string;
  difficulty: string;
  questionsCount: number;
  durationMinutes: number;
  attemptCount: number;
  isFree: boolean;
  /** @nullable */
  year?: number | null;
}