# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRI ANH EDU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: number;
  text: string;
  options: string[];
  correctOptionIndex: number;
  /** @nullable */
  explanation?: string | null;
}