# lib/api-zod/src/generated/types/teacher.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRI ANH EDU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Teacher {
  id: number;
  name: string;
  avatar: string;
  specialization: string;
  experienceYears: number;
  studentsCount: number;
  bio: string;
  /** @nullable */
  rating?: number | null;
}