# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TRI ANH EDU API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  studentName: string;
  studentAvatar: string;
  score: number;
  improvement: number;
  testimonialText: string;
  examType: string;
  /** @nullable */
  monthsStudied?: number | null;
}